Defer numba guvectorize compilation for conversion functions
